Moved TM APIs to Clone files
